feat: 感情取得 -> 抽選まで
